# Heap allocation optimizations