--- a/src/GraphQL.Benchmarks/GraphQL.Benchmarks.csproj
+++ b/src/GraphQL.Benchmarks/GraphQL.Benchmarks.csproj
@@ -1,13 +1,21 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
+    <OutputType>exe</OutputType>
     <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\GraphQL\GraphQL.csproj" />
     <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.0.0" />
   </ItemGroup>
-  
+
+  <ItemGroup>
+    <None Include="**\*.md" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GraphQL.StarWars\GraphQL.StarWars.csproj" />
+  </ItemGroup>
+
 </Project>