--- a/src/GraphQL.Benchmarks/Program.cs
+++ b/src/GraphQL.Benchmarks/Program.cs
@@ -20,7 +20,7 @@ namespace GraphQL.Benchmarks
             while (true)
             {
                 bench.Introspection();
-                Thread.Sleep(1000);
+                Thread.Sleep(100);
             }
         }
     }