--- a/src/GraphQL/EnumerableExtensions.cs
+++ b/src/GraphQL/EnumerableExtensions.cs
@@ -17,5 +17,18 @@ namespace GraphQL
                 action(item);
             }
         }
+
+        /// <summary>
+        /// Performs the indicated action on each item. Boxing free for <c>List+Enumerator{T}</c>.
+        /// </summary>
+        /// <param name="action">The action to be performed.</param>
+        /// <remarks>If an exception occurs, the action will not be performed on the remaining items.</remarks>
+        public static void Apply<T>(this List<T> items, Action<T> action)
+        {
+            foreach (var item in items)
+            {
+                action(item);
+            }
+        }
     }
 }