--- a/src/GraphQL/Execution/ExecutionNode.cs
+++ b/src/GraphQL/Execution/ExecutionNode.cs
@@ -124,7 +124,7 @@ namespace GraphQL.Execution
 
     public class ArrayExecutionNode : ExecutionNode, IParentExecutionNode
     {
-        public IList<ExecutionNode> Items { get; set; }
+        public List<ExecutionNode> Items { get; set; }
 
         public ArrayExecutionNode(ExecutionNode parent, IGraphType graphType, Field field, FieldType fieldDefinition, string[] path)
             : base(parent, graphType, field, fieldDefinition, path)