--- a/src/GraphQL/Execution/SubscriptionExecutionStrategy.cs
+++ b/src/GraphQL/Execution/SubscriptionExecutionStrategy.cs
@@ -109,20 +109,22 @@ namespace GraphQL.Execution
                     })
                     .SelectMany(async executionNode =>
                     {
-                        foreach (var listener in context.Listeners)
-                        {
-                            await listener.BeforeExecutionAsync(context.UserContext, context.CancellationToken)
-                                .ConfigureAwait(false);
-                        }
+                        if (context.Listeners != null)
+                            foreach (var listener in context.Listeners)
+                            {
+                                await listener.BeforeExecutionAsync(context.UserContext, context.CancellationToken)
+                                    .ConfigureAwait(false);
+                            }
 
                         // Execute the whole execution tree and return the result
                         await ExecuteNodeTreeAsync(context, executionNode).ConfigureAwait(false);
 
-                        foreach (var listener in context.Listeners)
-                        {
-                            await listener.AfterExecutionAsync(context.UserContext, context.CancellationToken)
-                                .ConfigureAwait(false);
-                        }
+                        if (context.Listeners != null)
+                            foreach (var listener in context.Listeners)
+                            {
+                                await listener.AfterExecutionAsync(context.UserContext, context.CancellationToken)
+                                    .ConfigureAwait(false);
+                            }
 
                         return new ExecutionResult
                         {