… appveyor.yml
… src/GraphQL.ApiTests/ApiApprovalTests.PublicApi.approved.txt
… src/GraphQL.Benchmarks/GraphQL.Benchmarks.Reference.md
… src/GraphQL/Introspection/__EnumValue.cs
… src/GraphQL/Introspection/__Field.cs
… src/GraphQL/Introspection/__Type.cs
… src/GraphQL/Language/AST/AbstractNode.cs
… src/GraphQL/Language/AST/Arguments.cs
… src/GraphQL/Language/AST/Directives.cs
… src/GraphQL/Language/AST/FragmentDefinition.cs
… src/GraphQL/Language/AST/FragmentSpread.cs
… src/GraphQL/Language/AST/Fragments.cs
… src/GraphQL/Language/AST/NullValue.cs
… src/GraphQL/Language/AST/Operation.cs
… src/GraphQL/Language/AST/SelectionSet.cs
… src/GraphQL/Language/AST/Variables.cs
… src/GraphQL/Language/CoreToVanillaConverter.cs
… src/GraphQL/TypeExtensions.cs
… src/GraphQL/Types/BooleanGraphType.cs
… src/GraphQL/Types/ComplexGraphType.cs
… src/GraphQL/Types/DirectiveGraphType.cs
… src/GraphQL/Types/EnumerationGraphType.cs
… src/GraphQL/Types/QueryArguments.cs
… src/GraphQL/Types/ResolveFieldContext.cs
… src/GraphQL/Utilities/AstPrinter.cs
… src/GraphQL/Utilities/EmptyEnumerator.cs
… src/GraphQL/Utilities/Federation/FederatedSchemaBuilder.cs
… src/GraphQL/Utilities/Federation/FederatedSchemaPrinter.cs
… src/GraphQL/Utilities/SchemaBuilder.cs
… src/GraphQL/Utilities/SchemaBuilderExtensions.cs