Introspection requests are good material for experimenting with memory allocation. This is what my test for requesting introspection shows:

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.14393.3243 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
Frequency=3515629 Hz, Resolution=284.4441 ns, Timer=TSC
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT  [AttachedDebugger]
  DefaultJob : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT


```
|        Method |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated | commit
|-------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|-------:|
| Introspection | 2.108 ms | 0.0406 ms | 0.0543 ms | 269.5313 | 93.7500 |     - |   1.25 MB | f0e192f [master] |
| Introspection | 1.910 ms | 0.0386 ms | 0.0814 ms | 230.4688 | 82.0313 |     - |   1.07 MB |  |
| Introspection | 1.661 ms | 0.0097 ms | 0.0086 ms | 222.6563 | 74.2188 |     - |   1.02 MB | d75a119 |
| Introspection | 1.567 ms | 0.0143 ms | 0.0134 ms | 222.6563 | 68.3594 |     - |   1.01 MB | c2e6b3c |
| Introspection | 1.568 ms | 0.0341 ms | 0.0615 ms | 246.0938 |       - |     - | 1020.74 KB | 51d31f3 |
| Introspection | 1.463 ms | 0.0132 ms | 0.0117 ms | 244.1406 | 7.8125 |     - |   1019 KB | b07256e |

This is just what I managed to fix, without really getting into code refactoring.

**[UPDATE]**
After #1500 :

|        Method |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated | commit
|-------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|-------:|
| Introspection | 1.470 ms | 0.0135 ms | 0.0126 ms | 179.6875 | 35.1563 |     - | 806.81 KB |