@sungam3r 2019-12-14
In all places indicated by this comment
`// DO NOT USE LINQ ON HOT PATH`
delegates actively allocated on the heap due to closures.