@sungam3r 2019-12-15
Fix for #1147 will be in a separate PR as it will require a design change.