@sungam3r 2019-12-15
@Shane32 I have some thoughts on `ResolveFieldContext` and your changes in #1441 
![image](https://user-images.githubusercontent.com/21261007/70871202-a5d19300-1fad-11ea-9307-f2209723de44.png)

`ResolveFieldContext` is the topmost