@Shane32 2019-12-16
@sungam3r Excellent; let's discuss.  I have also have some significant performance improvements to ResolveFieldContext I'd like to suggest, but they depend on the IFieldResolver interfaces being implemented.  I actually have the PR ready to go, which I could post now, but it depends on #1441 .  This new PR will cut the number of allocations by ExecutionStrategy per resolver run by as much as 80%, depending on the resolver, not to mention reduced memory requirements and increased execution speed.

I only glanced at this PR, but I see that one of the changes is to convert ResolveFieldContext to a struct.  If interfaces are used such as in #1441 , all performance enhancements (if any) are lost, as structs are boxed when cast to an interface.