@sungam3r 2020-01-11
@joemcbride @SimonCropp PR is ready for review. I resolved conflicts and double-checked each file. Many files have been changed, but all the changes come down to fairly simple operations that do not change the whole design. After merging this ASAP I would like to update parser dependency graphql-dotnet/parser#53 to see further effect.

@joemcbride I would appreciate your feedback. I would not want merging such PR without your approve.