@Shane32 2020-01-16
Looks good to me.  I think the boxing and short-circuit optimizations are making the real difference here.  LINQ optimizes calls to `Any()` and `Count()` (by testing if they can be cast before enumerating), and my benchmarks show that replacing IEnumerable with List has little effect.  However, every little bit helps, and the required changes to the API are minimal and seem fine to me.  Excellent work @sungam3r  !