@Shane32 2020-01-16
** If ** you wanted to break it down .... you could put the boxing optimizations in one PR, the short-circuit optimizations in another PR, and the removal of LINQ in a third PR.  Something like that.  Then it would be easier to review the code and/or run benchmarks.